router: separated stale channel handlers
